requirements.txt
# --extra-index-url https://pypi.org/simple
# fastapi==0.1.0
# uvicorn==0.27.1
# python-multipart==0.0.9
# pillow==9.1.0
# opencv-python-headless==4.5.5
# tensorflow==2.15.0
# numpy==1.24.3
# tensorflow-serving-api==2.14.1
fastapi==0.1.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==9.1.0
opencv-python-headless==4.5.5.64
tensorflow==2.16.1
numpy==1.24.3
tensorflow-serving-api==2.16.1